requirements.txt
python-dotenv==1.0.1
pydantic==2.7.4
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
openai==1.40.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
pandas==2.2.2
SQLAlchemy==2.0.32
streamlit==1.37.1
tqdm==4.66.4
seaborn>=0.13
networkx>=3.2
scikit-learn>=1.4
joblib>=1.3
python-docx>=1.1.2
PyYAML>=6.0.1
